+++ packages/perseus/src/user-input.types.ts
@@ -1,0 +1,14 @@
+export type UserInputStatus = "correct" | "incorrect" | "incomplete";
+
+export type PerseusCategorizerUserInput = {
+    values: ReadonlyArray<number>;
+};
+
+export type PerseusCSProgramUserInput = {
+    status: UserInputStatus;
+    message: string | null;
+};
+
+export type PerseusDropdownUserInput = {
+    value: number;
+};
