+++ packages/perseus/src/widgets/dropdown/dropdown-validator.test.tsx
@@ -1,12 +1,12 @@
 import dropdownValidator from "./dropdown-validator";
 import {question1} from "./dropdown.testdata";
 
-import type {UserInput} from "./dropdown.types";
+import type {PerseusDropdownUserInput} from "../../user-input.types";
 
 describe("dropdownValidator", () => {
     it("returns invalid for user input of 0", () => {
         // Arrange
-        const userInput: UserInput = {
+        const userInput: PerseusDropdownUserInput = {
             value: 0,
         };
         const rubric = question1.widgets["dropdown 1"].options;
@@ -23,7 +23,7 @@ describe("dropdownValidator", () => {
 
     it("returns 0 points for incorrect answer", () => {
         // Arrange
-        const userInput: UserInput = {
+        const userInput: PerseusDropdownUserInput = {
             value: 1,
         };
         const rubric = question1.widgets["dropdown 1"].options;
@@ -42,7 +42,7 @@ describe("dropdownValidator", () => {
 
     it("returns 1 point for correct answer", () => {
         // Arrange
-        const userInput: UserInput = {
+        const userInput: PerseusDropdownUserInput = {
             value: 2,
         };
         const rubric = question1.widgets["dropdown 1"].options;
